+++ src/Symfony/Component/Form/Extension/Core/Type/BaseType.php
@@ -13,7 +13,6 @@ namespace Symfony\Component\Form\Extension\Core\Type;
 
 use Symfony\Component\Form\AbstractRendererEngine;
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Exception\LogicException;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormView;
@@ -63,14 +62,6 @@ abstract class BaseType extends AbstractType
             if (!$labelFormat) {
                 $labelFormat = $view->parent->vars['label_format'];
             }
-
-            $rootFormAttrOption = $form->getRoot()->getConfig()->getOption('form_attr');
-            if ($options['form_attr'] || $rootFormAttrOption) {
-                $options['attr']['form'] = \is_string($rootFormAttrOption) ? $rootFormAttrOption : $form->getRoot()->getName();
-                if (empty($options['attr']['form'])) {
-                    throw new LogicException('"form_attr" option must be a string identifier on root form when it has no id.');
-                }
-            }
         } else {
             $id = \is_string($options['form_attr']) ? $options['form_attr'] : $name;
             $fullName = $name;
@@ -94,6 +85,9 @@ abstract class BaseType extends AbstractType
         $view->vars = array_replace($view->vars, [
             'form' => $view,
             'id' => $id,
+            // filled by FormType::finishView(), which is the first point where the "form_attr"
+            // option of every descendant is known
+            'form_id' => null,
             'name' => $name,
             'full_name' => $fullName,
             'disabled' => $form->isDisabled(),